Changing sschema for pgsql

+++ data/schema.sql
@@ -1,7 +1,7 @@
 drop table if exists cards;
 create table cards (
   id integer primary key,
-  type tinyint not null, /* 1 for vocab, 2 for code */
+  type INTEGER not null, /* 1 for vocab, 2 for code */
   front text not null,
   back text not null,
   known boolean default 0
